renamed custom to config and rearranged stuff
