+++ lua/config/lsp.lua
@@ -1,0 +1,7 @@
+return {
+    installed = { "rust_analyzer", "clangd", "omnisharp", "lua_ls" },
+    default_cfg = {},
+    custom_opts = {
+        rustfmt = {},
+    }
+}
