+++ lua/config/neovide.lua
